finetuning and evaluation/requirements.txt
numpy
pandas
matplotlib
scikit-learn
torch
transformers
google-genai
